# .docs/openapi.yaml
openapi: 3.0.3
info:
  title: Kegiatin API
  description: REST API for Kegiatin - Aplikasi Manajemen Kegiatan PD Pemuda Persis
  version: 1.0.0
  contact:
    name: Kelompok C6
servers:
  - url: http://localhost:3000/api
    description: Local Development Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EventType:
      type: string
      enum: [SINGLE, SERIES]
    EventStatus:
      type: string
      enum: [DRAFT, PUBLISHED, ONGOING, COMPLETED, CANCELLED]
    EventVisibility:
      type: string
      enum: [OPEN, INVITE_ONLY]
    SessionStatus:
      type: string
      enum: [SCHEDULED, ONGOING, COMPLETED, POSTPONED]
    RsvpStatus:
      type: string
      enum: [CONFIRMED, CANCELLED, WAITLIST]
...
    ApiResponse:
      type: object
      required: [success, statusCode]
      properties:
        success:
          type: boolean
        statusCode:
          type: integer
        message:
          type: string
        data:
          type: object
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer

    # ── Auth ──
...
    CreateEventDto:
      type: object
      required: [title, type, visibility, sessions]
      properties:
        title:
          type: string
        description:
          type: string
        type:
          $ref: "#/components/schemas/EventType"
        visibility:
          $ref: "#/components/schemas/EventVisibility"
        location:
          type: string
        contactPerson:
          type: string
          description: Nama dan kontak narahubung (e.g. "Pak Ahmad 08123456789")
        imageUrl:
          type: string
          nullable: true
          description: URL gambar/banner event (dari POST /uploads/image)
        maxParticipants:
          type: integer
          nullable: true
          description: Batas maksimal peserta RSVP (null = unlimited)
        sessions:
          type: array
          description: |
            Client selalu mengirimkan array sessions (minItems: 1).
            SINGLE event: kirim tepat 1 session.
            SERIES event: kirim N sessions sesuai pola jadwal yang dipilih user.
            Server tidak auto-generate session; client bertanggung jawab atas isi array ini.
          minItems: 1
          items:
            type: object
            required: [title, startTime, endTime]
            properties:
              title:
                type: string
              startTime:
                type: string
                format: date-time
              endTime:
                type: string
                format: date-time
              location:
                type: string
              capacity:
                type: integer
                nullable: true
...
    SessionResponse:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        location:
          type: string
        order:
          type: integer
        status:
          $ref: "#/components/schemas/SessionStatus"
        capacity:
          type: integer
          nullable: true
          description: null = unlimited
    EventResponse:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          $ref: "#/components/schemas/EventType"
        status:
          $ref: "#/components/schemas/EventStatus"
        visibility:
          $ref: "#/components/schemas/EventVisibility"
        location:
          type: string
        contactPerson:
          type: string
        imageUrl:
          type: string
          nullable: true
        maxParticipants:
          type: integer
          nullable: true
          description: Batas maksimal peserta RSVP (null = unlimited)
        createdBy:
          type: string
        sessions:
          type: array
          items:
            $ref: "#/components/schemas/SessionResponse"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ── Session ──
...
    RsvpResponse:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        eventId:
          type: string
        qrToken:
          type: string
          description: Custom algorithm token
        status:
          $ref: "#/components/schemas/RsvpStatus"
        createdAt:
          type: string
          format: date-time
    RsvpWithUserResponse:
      description: |
        Digunakan untuk endpoint GET /events/{id}/rsvp (Admin offline cache).
        Menyertakan data user agar admin dapat mengidentifikasi peserta tanpa koneksi.
      allOf:
        - $ref: "#/components/schemas/RsvpResponse"
        - type: object
          properties:
            user:
              type: object
              properties:
                displayName:
                  type: string
                npa:
                  type: string
                  nullable: true
                cabang:
                  type: string
                  nullable: true
                photoUrl:
                  type: string
                  nullable: true

    # ── Attendance ──
...
paths:
...
  /events:
    get:
      tags: [Events]
      summary: List events (paginated, filterable)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/EventStatus"
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/EventType"
        - name: visibility
          in: query
          schema:
            $ref: "#/components/schemas/EventVisibility"
        - name: search
          in: query
          description: Pencarian berdasarkan judul event (case-insensitive, partial match)
          schema:
            type: string
      responses:
        "200":
          description: Daftar event
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/EventResponse"
                      meta:
                        $ref: "#/components/schemas/PaginationMeta"
    post:
      tags: [Events]
      summary: Create event baru (Admin only)
      description: |
        Client selalu mengirimkan array sessions dalam request body (minItems: 1).
        SINGLE event: kirim tepat 1 session.
        SERIES event: kirim N sessions sesuai pola jadwal yang dipilih user.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventDto"
      responses:
        "201":
          description: Event berhasil dibuat (status DRAFT)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/EventResponse"
        "403":
          description: Bukan admin
...
  /events/{id}/rsvp:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      tags: [RSVP]
      summary: RSVP ke event (1 per user per event)
      description: |
        QR token di-generate server-side (custom algorithm).
        Untuk Series Event, 1 RSVP covers all sessions.
        Response includes qrToken untuk disimpan di Hive client.
      security:
        - bearerAuth: []
      responses:
        "201":
          description: RSVP berhasil, qrToken included
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/RsvpResponse"
        "400":
          description: Already RSVP'd or event not published
        "409":
          description: Duplicate RSVP
    get:
      tags: [RSVP]
      summary: List peserta RSVP per event (Admin, for offline cache)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
      responses:
        "200":
          description: Daftar peserta RSVP
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/RsvpWithUserResponse"
                      meta:
                        $ref: "#/components/schemas/PaginationMeta"
  /rsvp/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    delete:
      tags: [RSVP]
      summary: Cancel RSVP
      security:
        - bearerAuth: []
      responses:
        "200":
          description: RSVP cancelled
        "400":
          description: Cannot cancel after check-in
  /rsvp/{id}/qr:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [RSVP]
      summary: Get QR token (re-fetch jika cache hilang)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: QR token
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        type: object
                        properties:
                          qrToken:
                            type: string
  /rsvp/me:
    get:
      tags: [RSVP]
      summary: List RSVP milik user sendiri
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: My RSVPs
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/RsvpResponse"
                      meta:
                        $ref: "#/components/schemas/PaginationMeta"

  # ════════════ ATTENDANCE ════════════
...